requirements.txt
tree-sitter==0.20.4
tree-sitter-python==0.20.4
tree-sitter-java==0.20.2
rank-bm25==0.2.2
sentence-transformers==2.2.2
numpy>=1.24.0
boto3>=1.28.0
javalang==0.13.0
networkx>=3.1
opensearch-py>=2.4.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
